changed datatype for cabin and user

+++ src/main/java/com/bookingapp/cabin/backend/model/Cabin.java
@@ -18,14 +18,14 @@ public class Cabin {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cabin_id")
-    private Integer cabinId;
+    private Long cabinId;
 
 
     @Column(name = "cabin_name")
     private String cabinName;
 
 
-    public Cabin(Integer cabinId) {
+    public Cabin(Long cabinId) {
         this.cabinId = cabinId;
     }
